Add json response for user notes.
Fixes #1194

This work was done while pairing with @MattHeard.

This patch adds json response for user notes, in the form of
/user/:username/notes.json. There are two cases here, when the user
exists, and when the user doesn't.

When the user exists, rails will deal with the controller in the same
way, and only a view is needed. Note that an empty partial for
notes/user is also needed, as the @notes object has reference to the
user, which is needed when jsonify is doing its thing.

When the user doesn't exist, we need to add a explicit json response in
the controller, which returns an empty json object and a 404 HTTP
response.