Wrap longitude selection when editing a diary entry
This work was done while pairing with @hughdavenport

Fixes #1191.

As described by the bug report, if a user was editing a diary entry, and
zoomed out on the map, and clicked outside the nominal range of -180 to
180, then the longitude value was not being wrapped around to fit within
that range.

This patch adds a function to the javascript asset which wraps the value
to fit within that range. As the range is open at both ends (-180 and
180 are two separate values), an edge case was needed for that. The rest
of the range is done with a custom wrap function which makes use of the
mod function with an offset, and a custom sign function to keep the same
sign as entered.